+++ src/extensions/core-api/stores.ts
@@ -20,5 +20,3 @@
  */
 
 export { ExtensionStore } from "../extension-store";
-export { MainIpcStore } from "../main-ipc-store";
-export { RendererIpcStore } from "../renderer-ipc-store";
